Move to helpers; tweak docs
