+++ crates/ruff/src/rules/pylint/rules/yield_in_init.rs
@@ -2,13 +2,9 @@ use rustpython_parser::ast::Expr;
 
 use ruff_macros::{define_violation, derive_message_formats};
 
-use crate::ast::function_type;
-use crate::ast::function_type::FunctionType;
+use crate::rules::pylint::helpers::in_dunder_init;
 use crate::{
-    ast::types::{FunctionDef, Range, ScopeKind},
-    checkers::ast::Checker,
-    registry::Diagnostic,
-    violation::Violation,
+    ast::types::Range, checkers::ast::Checker, registry::Diagnostic, violation::Violation,
 };
 
 define_violation!(
@@ -50,36 +46,4 @@ pub fn yield_in_init(checker: &mut Checker, expr: &Expr) {
             .diagnostics
             .push(Diagnostic::new(YieldInInit, Range::from_located(expr)));
     }
-}
-
-pub fn in_dunder_init(checker: &mut Checker) -> bool {
-    let scope = checker.current_scope();
-    let ScopeKind::Function(FunctionDef {
-        name,
-        decorator_list,
-        ..
-    }) = &scope.kind else {
-        return false;
-    };
-    if *name != "__init__" {
-        return false;
-    }
-    let Some(parent) = checker.current_scope_parent() else {
-        return false;
-    };
-
-    if !matches!(
-        function_type::classify(
-            checker,
-            parent,
-            name,
-            decorator_list,
-            &checker.settings.pep8_naming.classmethod_decorators,
-            &checker.settings.pep8_naming.staticmethod_decorators,
-        ),
-        FunctionType::Method
-    ) {
-        return false;
-    }
-    true
 }
